# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Drifter Highway Space API",
        "description": "EVE Online authentication and character management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        },
        {
            "url": "https://drifterhighway.space",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": ["Users"],
                "summary": "Get All Users",
                "description": "Retrieve all active users from the database.",
                "operationId": "GetUsers",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Which page of results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - valid JWT token required"
                    },
                    "404": {
                        "description": "No users found"
                    }
                }
            },
            "put": {
                "tags": ["Users"],
                "summary": "Update User",
                "description": "Update a user by ID.",
                "operationId": "UpdateUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "User ID"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Users"],
                "summary": "Delete User",
                "description": "Delete (deactivate) a user by ID. User is marked as Inactive.",
                "operationId": "DeleteUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "User ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": ["Users"],
                "summary": "Get User by ID",
                "description": "Retrieve a specific user by ID.",
                "operationId": "GetUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "User ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": ["Users"],
                "summary": "Update User by ID",
                "description": "Update a specific user.",
                "operationId": "UpdateUserById",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "User ID"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Users"],
                "summary": "Delete User by ID",
                "description": "Delete (deactivate) a specific user.",
                "operationId": "DeleteUserById",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "User ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer Token for authentication"
            }
        },
        "schemas": {
...
            "User": {
                "type": "object",
                "description": "User entity from the database",
                "properties": {
                    "ID": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Unique identifier for the user"
                    },
                    "Name": {
                        "type": "string",
                        "description": "Name of the user"
                    },
                    "Status": {
                        "type": "string",
                        "description": "Status of the user (Active/Inactive)",
                        "enum": ["Active", "Inactive"]
                    },
                    "DiscordID": {
                        "type": "string",
                        "description": "Optional Discord ID for the user"
                    }
                }
            },
            "UserUpdate": {
                "type": "object",
                "description": "Request body for updating a user",
                "properties": {
                    "Name": {
                        "type": "string",
                        "description": "New name for the user"
                    },
                    "DiscordID": {
                        "type": "string",
                        "description": "Discord ID for the user"
                    }
                }
            },
...
        }
    }
}